moves Return Codes into other file
